+++ app/Http/Controllers/Admin/UserController.php
@@ -10,6 +10,7 @@ use App\Models\Permission;
 use App\Models\User;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Inertia\Inertia;
 
@@ -18,10 +19,12 @@ class UserController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+    * @return \Inertia\Response
      */
     public function index(Request $request)
     {
+        $this->authorize('users.viewAny', User::class);
+
         $result = User::search($request->term);
 
         return Inertia::render('User/Index', [
@@ -29,16 +32,22 @@ class UserController extends Controller
             'count' => $result['count'],
             'page' => $request->page?? 1,
             'termSearch' => $request->term,
+            'can' => [
+                'create' => Auth::user()->can('users.create'),
+                'view' => Auth::user()->can('users.view'),
+            ]
         ]);
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Inertia\Response
      */
     public function create()
     {
+        $this->authorize('users.create', User::class);
+
         return Inertia::render('User/Create', [
             'permissions' => Permission::select('id', 'description')->get(),
         ]);
@@ -46,12 +55,14 @@ class UserController extends Controller
 
     /**
      * Store a newly created resource in storage.
-     *
+     * Qwer@1234
      * @param  StoreUserRequest  $request
-     * @return Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(StoreUserRequest $request)
     {
+        $this->authorize('users.create', User::class);
+
         $data = $request->validated();
         $data['password'] = Hash::make($request->password);
 
@@ -59,7 +70,7 @@ class UserController extends Controller
             $user = User::create($data);
             return redirect()->route('users.show', $user)->with('flash', ['status' => 'success', 'message' => 'Registro salvo com sucesso.']);
         } catch (Exception $e) {
-            return redirect()->route('users.index')->with('flash', ['status' => 'danger', 'message' => $e->message]);
+            return redirect()->route('users.index')->with('flash', ['status' => 'danger', 'message' => $e->getMessage()]);
         }
     }
 
@@ -67,12 +78,18 @@ class UserController extends Controller
      * Display the specified resource.
      *
      * @param  User $user
-     * @return \Illuminate\Http\Response
+     * @return \Inertia\Response
      */
     public function show(User $user)
     {
+        $this->authorize('users.view', $user);
+
         return Inertia::render('User/Show', [
             'user' => User::with('permission')->find($user->id),
+            'can' => [
+                'update' => Auth::user()->can('users.update'),
+                'delete' => Auth::user()->can('users.delete'),
+            ]
         ]);
     }
 
@@ -80,10 +97,12 @@ class UserController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param  User $user
-     * @return \Illuminate\Http\Response
+     * @return \Inertia\Response
      */
     public function edit(User $user)
     {
+        $this->authorize('users.update', $user);
+
         return Inertia::render('User/Edit', [
             'user' => $user,
         ]);
@@ -93,10 +112,12 @@ class UserController extends Controller
      * Show the form for editing password the specified resource.
      *
      * @param  User $user
-     * @return \Illuminate\Http\Response
+     * @return \Inertia\Response
      */
     public function editPassword(User $user)
     {
+        $this->authorize('users.update.password', $user);
+
         return Inertia::render('User/EditPassword', [
             'user' => $user,
         ]);
@@ -107,17 +128,19 @@ class UserController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  User $user
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        $this->authorize('users.update', $user);
+
         $data = $request->validated();
 
         try {
             $user->update($data);
             return redirect()->route('users.show', $user)->with('flash', ['status' => 'success', 'message' => 'Registro atualizado com sucesso!']);
         } catch (Exception $e) {
-            return redirect()->route('users.index')->with('flash', ['status' => 'danger', 'message' => $e->message]);
+            return redirect()->route('users.index')->with('flash', ['status' => 'danger', 'message' => $e->getMessage()]);
         }
     }
 
@@ -126,17 +149,19 @@ class UserController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  User $user
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function updatePassword(UpdateUserPasswordRequest $request, User $user)
     {
+        $this->authorize('users.update.password', $user);
+
         $data = $request->validated();
 
         try {
             $user->update($data);
             return redirect()->route('users.show', $user)->with('flash', ['status' => 'success', 'message' => 'Registro atualizado com sucesso!']);
         } catch (Exception $e) {
-            return redirect()->route('users.index')->with('flash', ['status' => 'danger', 'message' => $e->message]);
+            return redirect()->route('users.index')->with('flash', ['status' => 'danger', 'message' => $e->getMessage()]);
         }
     }
 
@@ -144,15 +169,17 @@ class UserController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  User $user
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy(User $user)
     {
+        $this->authorize('users.delete', $user);
+
         try {
             $user->delete();
             return redirect()->route('users.index')->with('flash', ['status' => 'success', 'message' => 'Registro apagado com sucesso!']);
         } catch (Exception $e) {
-            return redirect()->route('users.index')->with('flash', ['status' => 'danger', 'message' => $e->message]);
+            return redirect()->route('users.index')->with('flash', ['status' => 'danger', 'message' => $e->getMessage()]);
         }
     }
 }
